replace bootstrap-application generator with app:bootstrap
